Version Packages (#14)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## emitnlog@0.2.0

### Minor Changes

- 766c97a: Retry option for the file operation for FileLogger and new
utilities (`createDeferredValue` and `startPolling`)
- 5c0fe92: Add notifier `OnEvent` type to make it easier to define
notifiable interfaces
- bc869df: Better support for non-blocking FileLogger and new utilities
(`delay` and `withTimeout`)
-   931eb57: New utility `isNotNullable`

### Patch Changes

-   6f0c479: Fix README Table of Contents for GitHub and NPM
-   1761fe9: README badges

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>