Version Packages (#23)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## emitnlog@0.3.0

### Minor Changes

- fbe25d9: Add `waitNextEvent` method to notifier allowing promise-based
listeners.
- efb7358: Adding `renew` to deferredValue (and also indicating if it
has been settled)

### Patch Changes

-   af99872: Exposing the test coverage as a badge

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>