Version Packages (#30)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## emitnlog@0.4.0

### Minor Changes

- 9d03081: Add Invocation Tracker: a lightweight utility to monitor
function calls, built on top of the emit n’ log core components.
- cfe5c30: PrefixedLogger: use `withPrefix` to automatically add
prefixes to log messages

### Patch Changes

-   e4c3f63: Rename logger/logger.ts to logger/definition.ts
-   13b4b5b: Document notifier 'waitForEvent'

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>