Version Packages (#43)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## emitnlog@0.5.0

### Minor Changes

- b9eb592: Add `mapOnEvent` to make it easier to expose internal
notifiers
- 8244da8: Add a dynamic version of `fromEnv` that supports both neutral
and node
-   78adc68: Debounce notifications and debounce utility
- 9f3b5fc: Add `fromEnv` to allow configuring the logger via environment
variables
-   be6a5f0: Add tags to `trackMethods`
- df3fb93: Add `PromiseTracker` and `trackPromise` for monitoring and
coordinating multiple unrelated promises
- 5a737c1: The tracker invocation now expose the 'stage' information as
a property
- c540029: Loggers now support a format option that includes a JSON
serialization version of a log line
-   d2ee3d7: Allow notifier without type parameter

### Patch Changes

-   c4996dc: Allow mapOnEvent to skip an event
-   28e902b: Add operation template to `trackMethods`

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>