Version Packages (#65)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## emitnlog@0.7.0

### Minor Changes

- 778b2f3: Add PromiseVault to help with code that require caching
promises.

### Patch Changes

-   bfebe80: Break README into smaller, more consumable, parts.
-   b2f6798: Improve stringify and add stringify options to loggers

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>