Version Packages (#104)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## emitnlog@0.9.1

### Patch Changes

- aca9e0d: Allow the `wait` method of `PromiseHolder` and `PromiseVault`
to wait for specific ids.
- 9f0e6e7: Improve the return types of the `closeAll` and `asClosable`
functions to better reflect whether the closables are synchronous or
asynchronous.

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>