Version Packages (#107)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## emitnlog@0.9.2

### Patch Changes

-   f9716d5: Add `withEmitLevel` to control the level of emitted entries
- b681bf1: enhance JSON formatting and add layout options to file logger
- 601ba3b: Add `withLevel` to create a logger with a different log level
- 5b2cfc4: Enhance `asExtendedLogger` and "with utils" to support prefix
loggers

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>