Version Packages (#116)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## emitnlog@0.10.0

### Minor Changes

- ce3a14b: LogEntry is now a main type on logger and only serializes
`args` if needed.
- e53b5af: New jsonParse utility (and helping types) to proper handle
deserialization of arbitrary objects

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>