Version Packages (#150)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## emitnlog@0.12.0

### Minor Changes

- f664528: Pass the option logger as a property of the input of
`runProcessMain`

### Patch Changes

-   76347a8: Fix `isProcessMain` for ESM

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>